fix: replace `sourceRoot` -> `rootDir` in tsconfig (#169)
This commit changes `sourceRoot` to `rootDir` in tsconfig.json for 4 of
the packages in this repo. (The other packages already used `rootDir`.)
This change is helpful because using `sourceRoot` creates sourcemaps
that some tools cannot read correctly, leading to build warnings and/or
debuggers like VSCode being unable to debug into original source files.